requirements.txt
Django==4.2.11
djangorestframework==3.14.0
django-filter==23.5
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.5.1
python-dotenv==1.0.1
dj-database-url==2.1.0
pymongo==4.6.3
python-dateutil==2.9.0.post0
gunicorn==21.2.0
django-cors-headers==4.4.0
psycopg[binary]==3.2.12
requests==2.32.3
whitenoise==6.7.0
tzdata==2025.1; platform_system == "Windows"

# Añadir:
celery>=5.2.7,<6.0
asgiref>=3.5.2,<4
# opcional (si usas estas apps):
django-celery-beat==2.6.0
django-celery-results==2.5.1